Add user interfaces with a basic console
There's no keybinding logic yet for this to be opened (for there is no keybind logic to begin with),
but hopefully I'll implement it before I forget about it.

The idea of this console is to be similar to that of the Source engine, where you can type debug
commands in order to change game variables.

This can probably also have output sent to it for debugging (perhaps by toggling a `debug` setting)